requirements.txt
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
openpyxl==3.1.2
requests==2.31.0
python-dotenv==1.0.0
airtable-python-wrapper==0.15.3
pinecone-client==3.0.0
scikit-learn==1.3.2
pytest==7.4.3
black==23.12.1
pyarrow
fastparquet